% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcce-package.R
\docType{package}
\name{stratselect-package}
\alias{stratselect}
\alias{stratselect-package}
\title{stratselect: Models for determining decision strategies in consumer choice experiments}
\description{
This package contains the likelihood functions necessary to recreate the modelling work for my PhD.
}
\author{
\strong{Maintainer}: Gavin Cooper \email{gavin@gavincooper.net} (\href{https://orcid.org/0000-0002-7186-5293}{ORCID})

}
\keyword{internal}
